+++ ci/cpu/cuspatial/build_cuspatial.sh
@@ -1,5 +1,0 @@
-set -e
-
-echo "Building cuspatial"
-conda build conda/recipes/cuspatial --python=$PYTHON
-